+++ packages/core/src/index.ts
@@ -37,5 +37,5 @@ export { SmartFilterOptimizer } from "./optimization/smartFilterOptimizer.js";
 export * from "./utils/textureLoaders.js";
 export * from "./serialization/index.js";
 export * from "./editorUtils/editableInPropertyPage.js";
-export * from "./utils/buildTools/shaderParsing.js";
+export * from "./utils/buildTools/shaderConverter.js";
 export * from "./serialization/importFragmentShader.js";
